Remove rides index page
There's no link to this page on the site anywhere that I can see, but
there have been 56 logged attempts to GET it in the past week. This page
doesn't really make sense for the application anyway, so this change
just has /rides redirect to the root. If a user is signed in they'll be
taken to their dashboard (which has a list of rides), otherwise they'll
get the getting started page.